+++ runtimes/haskell/lbr-plutarch/src/LambdaBuffers/Runtime/Plutarch/LamVal.hs
@@ -18,17 +18,15 @@ module LambdaBuffers.Runtime.Plutarch.LamVal (
 
 import Plutarch.Builtin.Data (
   PBuiltinList (PNil),
+  PBuiltinPair (PBuiltinPair),
   PData,
   pasConstr,
   pasInt,
   pasList,
   pchooseData,
   pconstrBuiltin,
-  pfstBuiltin,
-  psndBuiltin,
  )
 import Plutarch.Prelude (
-  ClosedTerm,
   PAsData,
   PBuiltinList (PCons),
   PInteger,
@@ -42,7 +40,7 @@ import Plutarch.Prelude (
   pforgetData,
   phoistAcyclic,
   plam,
-  plet,
+  pmatch,
   ptraceInfo,
   ptryFrom,
   (#),
@@ -51,7 +49,7 @@ import Plutarch.Prelude (
 import Plutarch.Unsafe (punsafeCoerce)
 
 -- | Plutarch `toPlutusData :: a -> PlutusData`
-ptoPlutusData :: ClosedTerm (PAsData a :--> PData)
+ptoPlutusData :: Term s (PAsData a :--> PData)
 ptoPlutusData = plam toPlutusData
 
 -- | Haskell `toPlutusData :: a -> PlutusData`
@@ -59,18 +57,18 @@ toPlutusData :: Term s (PAsData a) -> Term s PData
 toPlutusData = pforgetData
 
 -- | Plutarch PlutusType `fromPlutusData :: PlutusData -> Parser a`
-pfromPlutusDataPlutusType :: ClosedTerm (PData :--> PAsData a)
+pfromPlutusDataPlutusType :: Term s (PData :--> PAsData a)
 pfromPlutusDataPlutusType = plam punsafeCoerce
 
 -- | Plutarch PTryFrom `fromPlutusData :: PlutusData -> Parser a`
-pfromPlutusDataPTryFrom :: PTryFrom PData (PAsData a) => ClosedTerm (PData :--> PAsData a)
+pfromPlutusDataPTryFrom :: PTryFrom PData (PAsData a) => Term s (PData :--> PAsData a)
 pfromPlutusDataPTryFrom = phoistAcyclic $ plam ptryFromData
   where
     ptryFromData :: forall a s. PTryFrom PData (PAsData a) => Term s PData -> Term s (PAsData a)
     ptryFromData pd = ptryFrom @(PAsData a) pd fst
 
 -- | Plutarch `constrData :: IntE -> ListE PlutusData -> PlutusData`
-pconstrData :: ClosedTerm (PInteger :--> PBuiltinList PData :--> PData)
+pconstrData :: Term s (PInteger :--> PBuiltinList PData :--> PData)
 pconstrData = phoistAcyclic $ plam $ \ix args -> pforgetData $ pconstrBuiltin # ix # args
 
 -- | Haskell `constrData :: IntE -> ListE PlutusData -> PlutusData`
@@ -78,7 +76,7 @@ constrData :: Term s PInteger -> [Term s PData] -> Term s PData
 constrData ix args = pforgetData $ pconstrBuiltin # ix # toBuiltinList args
 
 -- | Plutarch `integerData :: IntE -> PlutusData`
-pintegerData :: ClosedTerm (PInteger :--> PData)
+pintegerData :: Term s (PInteger :--> PData)
 pintegerData = phoistAcyclic $ plam $ \i -> ptoPlutusData # pdata i
 
 -- | Haskell `integerData :: IntE -> PlutusData`
@@ -86,7 +84,7 @@ integerData :: Term s PInteger -> Term s PData
 integerData = toPlutusData . pdata
 
 -- | Plutarch `listData :: ListE PlutusData -> PlutusData`
-plistData :: ClosedTerm (PBuiltinList PData :--> PData)
+plistData :: Term s (PBuiltinList PData :--> PData)
 plistData = phoistAcyclic $ plam $ pforgetData . pdata
 
 -- | Haskell `listData :: ListE PlutusData -> PlutusData`
@@ -99,12 +97,12 @@ toBuiltinList (x : xs) = pcon (PCons x (toBuiltinList xs))
 
 -- | Plutarch `casePlutusData :: (Int -> [PlutusData] -> a) -> ([PlutusData] -> a) -> (Int -> a) -> (PlutusData -> a) -> PlutusData -> a`
 pcasePlutusData ::
-  ClosedTerm ((PInteger :--> PBuiltinList PData :--> a) :--> (PBuiltinList PData :--> a) :--> (PInteger :--> a) :--> (PData :--> a) :--> PData :--> a)
+  Term s ((PInteger :--> PBuiltinList PData :--> a) :--> (PBuiltinList PData :--> a) :--> (PInteger :--> a) :--> (PData :--> a) :--> PData :--> a)
 pcasePlutusData = phoistAcyclic $ plam $ \handleConstr handleList handleInt handleOther pd ->
   pforce $
     pchooseData
       # pd
-      # pdelay (plet (pasConstr # pd) $ \pair -> handleConstr # (pfstBuiltin # pair) # (psndBuiltin # pair))
+      # pdelay (pmatch (pasConstr # pd) $ \(PBuiltinPair ix args) -> handleConstr # ix # args)
       # pdelay (ptraceInfo "Got a PlutusData Map" (handleOther # pd))
       # pdelay (handleList # (pasList # pd))
       # pdelay (handleInt # (pasInt # pd))
@@ -121,13 +119,13 @@ casePlutusData ::
 casePlutusData handleConstr handleList handleInt handleOther pd = pcasePlutusData # plam handleConstr # plam handleList # plam handleInt # plam handleOther # pd
 
 -- | Plutarch `succeedParse :: a -> Parser a`
-psucceedParse :: ClosedTerm (a :--> a)
+psucceedParse :: Term s (a :--> a)
 psucceedParse = plam id
 
 -- | Plutarch `failParse :: Parser a`
-pfailParse :: ClosedTerm a
+pfailParse :: Term s a
 pfailParse = perror
 
 -- | Plutarch `bindParse :: Parser a -> (a -> Parser b) -> Parser b`
-pbindParse :: ClosedTerm (a :--> (a :--> b) :--> b)
+pbindParse :: Term s (a :--> (a :--> b) :--> b)
 pbindParse = phoistAcyclic $ plam (flip (#))